Code cleanup and directory fix
