+++ src/js/index.js
@@ -1,0 +1,5 @@
+import 'bootstrap/dist/css/bootstrap.min.css';
+import '../css/main.css';
+import addEventListeners from './eventListeners';
+
+addEventListeners();
